+++ modules/manager/run.go
@@ -28,7 +28,9 @@ import (
 	"github.com/samber/lo"
 	corev1 "k8s.io/api/core/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/selection"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -36,6 +38,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/config"
 	"sigs.k8s.io/controller-runtime/pkg/healthz"
+
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/metrics/filters"
@@ -79,6 +82,10 @@ type Config struct {
 	EnforceConfig            bool
 	ClusterDomain            string
 	EmitKubernetesEvents     bool
+	// SecretLabelSelector specifies the label which will be used to limit the ingestion of secrets. Only those that have this label set to "true" will be ingested.
+	SecretLabelSelector string
+	// ConfigMapLabelSelector specifies the label which will be used to limit the ingestion of configmaps. Only those that have this label set to "true" will be ingested.
+	ConfigMapLabelSelector string
 
 	// ServiceAccountToImpersonate is the name of the service account to impersonate,
 	// by the controller manager, when making requests to the API server.
@@ -119,6 +126,8 @@ func DefaultConfig() Config {
 		ClusterCASecretName:           "kong-operator-ca",
 		ClusterCASecretNamespace:      defaultNamespace,
 		ControllerNamespace:           defaultNamespace,
+		SecretLabelSelector:           mgrconfig.DefaultSecretLabelSelector,
+		ConfigMapLabelSelector:        mgrconfig.DefaultConfigMapLabelSelector,
 		LoggerOpts:                    &zap.Options{},
 		GatewayControllerEnabled:      true,
 		ControlPlaneControllerEnabled: true,
@@ -184,6 +193,27 @@ func Run(
 	if cfg.CacheSyncPeriod > 0 {
 		setupLog.Info("cache sync period set", "period", cfg.CacheSyncPeriod)
 		cacheOptions.SyncPeriod = &cfg.CacheSyncPeriod
+	}
+
+	cacheOptions.ByObject = map[client.Object]cache.ByObject{}
+	if cfg.SecretLabelSelector != "" {
+		req, err := labels.NewRequirement(cfg.SecretLabelSelector, selection.Equals, []string{"true"})
+		if err != nil {
+			return fmt.Errorf("failed to make label requirement for secrets: %w", err)
+		}
+		cacheOptions.ByObject[&corev1.Secret{}] = cache.ByObject{
+			Label: labels.NewSelector().Add(*req),
+		}
+	}
+
+	if cfg.ConfigMapLabelSelector != "" {
+		req, err := labels.NewRequirement(cfg.ConfigMapLabelSelector, selection.Equals, []string{"true"})
+		if err != nil {
+			return fmt.Errorf("failed to make label requirement for config maps: %w", err)
+		}
+		cacheOptions.ByObject[&corev1.ConfigMap{}] = cache.ByObject{
+			Label: labels.NewSelector().Add(*req),
+		}
 	}
 
 	mgr, err := ctrl.NewManager(