+++ CHANGELOG.md
@@ -60,6 +60,15 @@
   This is a breaking change, as any code or configuration relying on the previous default will now use 'cluster.local'
   unless explicitly overridden.
   [#1870](https://github.com/Kong/kong-operator/pull/1870)
+- Add flags `--secret-label-selector` and `--config-map-label-selector` to
+  filter watched `Secret`s and `ConfigMap`s. Only secrets or configMaps with
+  the given label and has the value set to `true` are reconciled by the
+  controllers. For example, if `--secret-label-selector` is set to `konghq.com/secret`,
+  only `Secret`s with the label `konghq.com/secret:true` are reconciled.
+  The default value of the two labels are set to `konghq.com/secret` and
+  `konghq.com/configmap` so it breaks when you are using `Secret`s and
+  `ConfigMap`s without the given labels.
+  [#1922](https://github.com/Kong/kong-operator/pull/1922)
 
 ### Added